removed placeholder files created by creation template
