--- docs/ref/modules/saltext.grafana.modules.grafana_mod.rst
@@ -1,5 +1,0 @@
-``grafana``
-===========
-
-.. automodule:: saltext.grafana.modules.grafana_mod
-    :members:
