--- docs/ref/modules/saltext.grafana.modules.rst
@@ -1,6 +1,0 @@
-
-saltext.grafana.modules.grafana_mod
-===================================
-
-.. automodule:: saltext.grafana.modules.grafana_mod
-    :members:
