--- docs/ref/states/saltext.grafana.states.grafana_mod.rst
@@ -1,5 +1,0 @@
-``grafana``
-===========
-
-.. automodule:: saltext.grafana.states.grafana_mod
-    :members:
