--- docs/ref/states/saltext.grafana.states.rst
@@ -1,6 +1,0 @@
-
-saltext.grafana.states.grafana_mod
-==================================
-
-.. automodule:: saltext.grafana.states.grafana_mod
-    :members:
