--- tests/integration/modules/test_grafana.py
@@ -1,13 +1,0 @@
-import pytest
-
-pytestmark = [
-    pytest.mark.requires_salt_modules("grafana.example_function"),
-]
-
-
-def test_replace_this_this_with_something_meaningful(salt_call_cli):
-    echo_str = "Echoed!"
-    ret = salt_call_cli.run("grafana.example_function", echo_str)
-    assert ret.exitcode == 0
-    assert ret.json
-    assert ret.json == echo_str
